backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.7
langgraph==0.2.45
faiss-cpu==1.9.0
python-dotenv==1.0.1
pydantic==2.9.2
